Apply Bias Correction to Large Ensemble Forecast Data Sets

This wrapper allows to apply the bias correction methods supplied with debias to large arrays of ensemble forecast and observation data.

debiasApply(fcst, obs, method = "unbias", fcst.out = fcst, ...,
  atomic = FALSE)

Arguments

fcst
array of ensemble forecast data with arbitrary dimensions. The final dimensions contain lead times, forecast instances (usually years) and ensemble members in this order.

obs
array of observation used for calibration. Dimensions correspond to the fcst array without the last dimension.

method
character string with bias correction method name

fcst.out
array of ensemble forecast data that is to be calibrated. All but the forecast instance and ensemble dimensions have to be the same as for fcst.

...
Additional parameters passed on to debias

atomic
logical, should debiasing be applied to each forecast instance separately (in contrast to jointly over lead times)?