List Calibration Methods

list_methods lists calibration methods available in biascorrection and print.biascorrection is used to nicely format these for display.

list_methods()

# S3 method for biascorrection
print(x, ...)

Arguments

x
data frame of methods to be used with debias

...
additional parameters for consistency with generic method

Examples

list_methods()

#>  METHODS          DESCRIPTION                                   
#>  ccr              Climate Conserving Recalibration              
#>  ccrlm            Climate Conserving Recalibration (as          
#>                   Regression)                                   
#>  comb             Conditional Bias With Linear Time Trend       
#>  conditional      Bias Conditional on Forecast                  
#>  initial          Bias dependent on initial condition           
#>  linmod           Linear Models for Bias Correction             
#>  monthly          Daily Calibration with Monthly Mean           
#>  moving           Moving Window Mean De-biasing                 
#>  mul              Multiplicative De-biasing                     
#>  qqmap            Quantile Mapping                              
#>  smooth           Mean De-biasing With Smoothing of Daily       
#>                   Climatology                                   
#>  smooth_mul       Multiplicative De-biasing With Smoothed       
#>                   Climatologies                                 
#>  smooth_scale     Smoothed Mean De-biasing With Variance Scaling
#>  smoothobs_mul    Multiplicative De-biasing With Smoothed       
#>                   Observation Climatology                       
#>  smoothobs_scale  Mean De-biasing With Variance Scaling         
#>  trend            Bias With Linear Time Trend                   
#>  useqmap          Quantile Mapping Using the 'qmap' Package